Moves speech related modules into dedicated subfolder.
