+++ ts/a11y/explorer/Region.ts
@@ -26,7 +26,7 @@
 import {MathDocument} from '../../core/MathDocument.js';
 import {CssStyles} from '../../util/StyleList.js';
 import {Sre} from '../sre.js';
-import {SsmlElement, buildSpeech} from '../SpeechUtil.js';
+import {SsmlElement, buildSpeech} from '../speech/SpeechUtil.js';
 
 export type A11yDocument = MathDocument<HTMLElement, Text, Document>;
