+++ ts/a11y/speech/SpeechMenu.ts
@@ -21,11 +21,11 @@
  * @author v.sorge@mathjax.org (Volker Sorge)
  */
 
-import { SpeechExplorer } from './explorer/KeyExplorer.js';
-import { ExplorerMathItem } from './explorer.js';
-import {MJContextMenu} from '../ui/menu/MJContextMenu.js';
-import {SubMenu, Submenu} from '../ui/menu/mj-context-menu.js';
-import {Sre} from './sre.js';
+import { SpeechExplorer } from '../explorer/KeyExplorer.js';
+import { ExplorerMathItem } from '../explorer.js';
+import {MJContextMenu} from '../../ui/menu/MJContextMenu.js';
+import {SubMenu, Submenu} from '../../ui/menu/mj-context-menu.js';
+import {Sre} from '../sre.js';
 
 /**
  * Values for the ClearSpeak preference variables.
